Split selectors.js into individual files
