+++ shared/lib/themes/selectors/current-theme.js
@@ -1,0 +1,11 @@
+export function getCurrentTheme( state, siteId ) {
+	return state.themes.currentTheme.get( 'currentThemes' ).get( siteId );
+}
+
+export function isActivating( state ) {
+	return state.themes.currentTheme.get( 'isActivating' );
+}
+
+export function hasActivated( state ) {
+	return state.themes.currentTheme.get( 'hasActivated' );
+}
